Refactor run simulation script
The script will now copy the example that is shipped with the container into the scripts folder in an example folder.
Additionally, the behaviour of the script when searching for external scripts was improved.
Scripts that are placed in the baselevel of the mounted scripts folder will be prioritized.
If no scripts are found the scripts in the examples folder will be used for the simulation.
In case that a specific script name is provided it will only be executed if the specified script was found.